+++ valhalla/__init__.py
@@ -4,16 +4,23 @@ from functools import partial
 from inspect import getmembers
 
 from filters import lookup
+from filters.casting import none
 
 class ValidationError(Exception): pass
 
 class Schema(object):
 
-	def __init__(self, name, match={}):
+	def __init__(self, name, match=[], require=[], blank=[], extra='ignore'):
 		self._name = name
 		self._fields = {}
 		self._valid = False
-		self._match = match
+		self._extra = extra
+		self._missing = []
+
+		self._field_options = {}
+		self._field_options['match'] = match
+		self._field_options['require'] = require
+		self._field_options['blank'] = blank
 
 	def __getattr__(self, attr):
 		return self.add_field(attr, Field(self, attr))
@@ -29,6 +36,10 @@ class Schema(object):
 	def valid(self):
 		return self._valid
 
+	@property
+	def missing(self):
+		return list(self._missing)
+
 	def add_field(self, name, field):
 		self._fields[name] = field
 		setattr(self, name, field)
@@ -39,19 +50,47 @@ class Schema(object):
 		for name, field in self._fields.iteritems():
 			if not field.validate(data_dict.get(name)):
 				self._valid = False
+
+	# options are ['all', None, [], ['fields']]
+	def _process_required(self, data_dict):
+		if not self._field_options['require']:
+			return
 		
-		for f1_name, f2_name in self._match.iteritems():
+		if self._field_options['require'] == 'all':
+			self._field_options['require'] = self._fields.keys()
+		
+		for f_name in self._field_options['require']:
 			try:
-				f1, f2 = self._fields[f1_name], self._fields[f2_name]
-			except KeyError:
-				continue
-			
-			if f1.result != f2.result:
+				field = self._fields[f_name]
+				field.require()
+			except KeyError: pass
+
+		required_fields = set([f.name for f in self._fields if f.required])
+		supplied_fields = set(data_dict.keys())
+
+		missing_fields = required_fields - supplied_fields
+		for f_name in missing_fields:
+			field = self._fields[f_name]
+			field._valid = False
+			field._errors.append('This field cannot be missing.')
+
+	def _process_matching(self, data_dict):
+		if not self._field_options['match']:
+			return
+
+		for f_names in self._field_options['match']:
+			fields = _names_to_fields(f_names)
+			if not all(f.result == fields[0].result for f in fields):
 				self._valid = False
-				f1._valid = False
-				f2._valid = False
-				f1._errors.append('This field must match %s' % f2)
-				f2._errors.append('This field must match %s' % f1)
+				self._errors.append('Fields [%s] do not match.' % ' '.join(f_names))
+
+	def _names_to_fields(field_names):
+		fields = []
+		for f_name in field_names:
+			try:
+				fields.append(self._fields[f_name])
+			except KeyError: pass
+		return fields
 
 
 	def reset(self):
@@ -63,16 +102,14 @@ class Field(object):
 	def __init__(self, schema, name):
 		self._schema = schema
 		self._name = name
-		self._original_value = None
-		self._value = None
 		self._filters = []
-		self._valid = False
-		self._errors = []
-
-		self._required = False
+		self._required = None
+		self._blank = None
 		self._alternate_name = None
+		self._match = None
 
-		self._ran = False
+		self.reset()
+
 
 	def __repr__(self):
 		return '[Field: %r] - [%r, %r, %r] - %r' % (self.name, self._original_value, self._value, self._valid, self._errors)
@@ -97,6 +134,18 @@ class Field(object):
 	def result(self):
 		return self._value
 
+	@property
+	def required(self):
+		if self._required is None:
+			return False
+		return self._required
+
+	@property
+	def blank_allowed(self):
+		if self._blank is None:
+			return True
+		return self._blank
+
 	def reset(self):
 		self._ran = False
 		self._valid = False
@@ -105,40 +154,45 @@ class Field(object):
 
 	def validate(self, value):
 		if self._ran:
-			return
+			return self._valid
 
-		self._original_value = self._value = value
-		
-		try:
-			for f in self._filters:
-				self._value = f.run(self._value)
-		except ValidationError as e:
+		self._original_value = self._value = none(_value=value)
+
+		if not self.blank_allowed and self._value is None:
 			self._valid = False
-			self._errors.append(e.message)
+			self._errors.append('This field cannot be blank.')
 		else:
-			self._valid = True
+			try:
+				for f in self._filters:
+					self._value = f.run(self._value)
+			except ValidationError as e:
+				self._valid = False
+				self._errors.append(e.message)
+			else:
+				self._valid = True
 
-		self._check_options()
 		return self._valid
 
-	def _check_options(self):
-		if self._value is None and self._required:
-			self._errors.append('This field is required.')
-			self._valid = False
+	def require(self, value=True):
+		if self._required is None:
+			self._required = value
+		return self
+
+	def blank(self, value=True):
+		if self._blank is None:
+			self._blank = value
+		return self
+
+	def alt(self, alt_name):
+		self._schema.add_field(alt_name, self)
+		return self
+
+	def match(self, *other_fields):
+		if not self._match:
+			self._match = other_fields
+		return self
 
 	def __call__(self, *args, **kwargs):
-		option_alias_map = {
-			'req': '_required',
-			'alt': '_alternate_name',
-		}
-	
-		for k, v in kwargs.iteritems():
-			if k in option_alias_map:
-				setattr(self, option_alias_map[k], v)
-
-		if self._alternate_name:
-			self._schema.add_field(self._alternate_name, self)
-
 		return self
 
 	def __getattr__(self, name):