+++ valhalla/tests/__init__.py
@@ -1,1 +1,6 @@
 # -*- coding: utf-8 -*-
+
+from .. import Schema
+
+def _schema():
+	return Schema('Test Schema')