+++ valhalla/tests/test_strings.py
@@ -2,10 +2,7 @@
 
 from nose.tools.trivial import assert_equals, assert_true, assert_false
 
-from .. import Schema
-
-def _blank_schema():
-	return Schema('Sample Schema')
+from . import _schema
 
 def _sample_data():
 	return {
@@ -25,26 +22,26 @@ def _sample_data():
 	}
 
 def test_text():
-	s = _blank_schema()
+	s = _schema()
 	s.random_string.text()
 	s.validate(_sample_data())
 
 	assert_true(s.valid)
 
-	s = _blank_schema()
+	s = _schema()
 	s.actual_numbers.text()
 
 	s.validate(_sample_data())
 	assert_false(s.valid)
 
 def test_alnum():
-	s = _blank_schema()
+	s = _schema()
 	s.alnum_string.alnum()
 	s.validate(_sample_data())
 
 	assert_true(s.valid)
 
-	s = _blank_schema()
+	s = _schema()
 	s.alpha_string.alnum()
 	s.actual_numbers.alnum()
 
@@ -55,13 +52,13 @@ def test_alnum():
 	assert_true(s.alpha_string.valid)
 
 def test_alpha():
-	s = _blank_schema()
+	s = _schema()
 	s.alpha_string.alpha()
 	s.validate(_sample_data())
 
 	assert_true(s.valid)
 
-	s = _blank_schema()
+	s = _schema()
 	s.actual_numbers.alpha()
 	s.alnum_string.alpha()
 
@@ -70,13 +67,13 @@ def test_alpha():
 	assert_false(s.alnum_string.valid)
 
 def test_numeric_string():
-	s = _blank_schema()
+	s = _schema()
 	s.numeric_string.numeric_string()
 	s.validate(_sample_data())
 
 	assert_true(s.valid)
 
-	s = _blank_schema()
+	s = _schema()
 
 	s.actual_numbers.numeric_string()
 	s.validate(_sample_data())
@@ -84,20 +81,20 @@ def test_numeric_string():
 	assert_false(s.valid)
 
 def test_nonblank():
-	s = _blank_schema()
+	s = _schema()
 	s.random_string.nonblank()
 	s.validate(_sample_data())
 
 	assert_true(s.valid)
 
-	s = _blank_schema()
+	s = _schema()
 	s.blank_string.nonblank()
 	s.validate(_sample_data())
 
 	assert_false(s.valid)
 
 def test_removespaces():
-	s = _blank_schema()
+	s = _schema()
 	s.awkward_spaces.removespaces()
 	s.validate(_sample_data())
 
@@ -105,7 +102,7 @@ def test_removespaces():
 	assert_equals(s.awkward_spaces.result, 'openthedoorgetonthefloorrr')
 
 def test_strip():
-	s = _blank_schema()
+	s = _schema()
 	s.flanking_spaces.strip()
 	s.validate(_sample_data())
 
@@ -113,7 +110,7 @@ def test_strip():
 	assert_equals(s.flanking_spaces.result, 'everybody walk the dinosaur')
 
 def test_lower():
-	s = _blank_schema()
+	s = _schema()
 	s.upper_case.lower()
 	s.validate(_sample_data())
 
@@ -121,7 +118,7 @@ def test_lower():
 	assert_equals(s.upper_case.result, 'i am the rawrmachine')
 
 def test_upper():
-	s = _blank_schema()
+	s = _schema()
 	s.lower_case.upper()
 	s.validate(_sample_data())
 
@@ -129,20 +126,20 @@ def test_upper():
 	assert_equals(s.lower_case.result, 'I AM THE QUIET MOUSE')
 
 def test_regex():
-	s = _blank_schema()
+	s = _schema()
 	s.numeric_string.regex(r'\d')
 	s.validate(_sample_data())
 
 	assert_true(s.valid)
 
-	s = _blank_schema()
+	s = _schema()
 	s.alpha_string.regex(r'\d')
 	s.validate(_sample_data())
 
 	assert_false(s.valid)
 
 def test_canonize():
-	s = _blank_schema()
+	s = _schema()
 	s.noncanonical_string.canonize()
 
 	s.validate(_sample_data())
@@ -151,7 +148,7 @@ def test_canonize():
 	assert_equals(s.noncanonical_string.result, u'there_is_nothing_canonical_aboutmeeeeee333_see?')
 
 def test_slugify():
-	s = _blank_schema()
+	s = _schema()
 	s.notslugified.slugify()
 
 	s.validate(_sample_data())